refactor: move test endpoints to credential controllers
- Move Notion /test endpoint from NotionController to NotionCredentialController
- Move Toggl /test endpoint from ToggleTrackController to TogglCredentialController
- Delete unused NotionController.java and ToggleTrackController.java files
- Update frontend API client paths:
  - /api/external/notion/test → /api/credentials/notion/test
  - /api/external/toggl/test → /api/credentials/toggl/test
- Consolidate all credential-related functionality in respective credential controllers
