+++ backend/src/main/java/com/example/backend/presentation/controllers/credentials/notion/NotionCredentialController.java
@@ -4,6 +4,7 @@ import com.example.backend.application.dto.credentials.notion.NotionCredentialCr
 import com.example.backend.application.dto.credentials.notion.NotionCredentialResponseDto;
 import com.example.backend.application.dto.credentials.notion.NotionCredentialUpdateRequestDto;
 import com.example.backend.application.usecases.credentials.notion.NotionCredentialUseCase;
+import com.example.backend.application.usecases.external.notion.NotionUseCase;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -26,6 +27,7 @@ import java.util.UUID;
 public class NotionCredentialController {
     
     private final NotionCredentialUseCase notionCredentialUseCase;
+    private final NotionUseCase notionUseCase;
     
     @PostMapping
     @Operation(summary = "Notion認証情報作成", description = "新しいNotion認証情報を作成します")
@@ -134,5 +136,16 @@ public class NotionCredentialController {
         
         boolean exists = notionCredentialUseCase.existsByUserId(userId);
         return ResponseEntity.ok(exists);
+    }
+    
+    @GetMapping("/test")
+    @Operation(summary = "Notion接続テスト", description = "Notion API接続をテストして、アクセス権限を確認します")
+    @ApiResponses({
+        @ApiResponse(responseCode = "200", description = "接続テストが正常に完了"),
+        @ApiResponse(responseCode = "500", description = "内部サーバーエラーまたはNotion API接続失敗")
+    })
+    public ResponseEntity<Boolean> testConnection() {
+        boolean isConnected = notionUseCase.testConnection();
+        return ResponseEntity.ok(isConnected);
     }
 }
